+++ seqspec/seqspec_index.py
@@ -1,9 +1,10 @@
-from seqspec.utils import load_spec, get_cuts
+from seqspec.utils import load_spec, map_read_id_to_regions
 from seqspec.seqspec_find import run_find
 from collections import defaultdict
 from typing import Dict, List, Tuple
 from argparse import SUPPRESS
 import os
+from seqspec.Region import RegionCoordinate, project_regions_to_coordinates, itx_read
 
 
 def setup_index_args(parser):
@@ -43,6 +44,12 @@ def setup_index_args(parser):
         "--rev", help="Returns 3'->5' region order", action="store_true"
     )
 
+    # boolean to indicate specifying a region or a read
+    subparser.add_argument(
+        "--region",
+        help="Specify a region",
+        action="store_true",
+    )
     subparser_required.add_argument(
         "-m",
         metavar="MODALITY",
@@ -58,13 +65,6 @@ def setup_index_args(parser):
         type=str,
         default=None,
         required=True,
-    )
-
-    # boolean to indicate specifying a region or a read
-    subparser.add_argument(
-        "--region",
-        help="Specify a region",
-        action="store_true",
     )
 
     return subparser
@@ -118,11 +118,12 @@ def run_index(
         if region:
             index = get_index(spec, modality, r, rev=rev)
         else:
-            index = get_index_by_primer(spec, modality, r, rev=rev)
-        indices.append({r: index})
+            index = get_index_by_primer(spec, modality, r)
+        indices.append(index)
     return FORMAT[fmt](indices, subregion_type)
 
 
+# TODO: modify to use RegionCoordinate object
 def get_index_by_type(
     spec, modality, region_id, rev=False
 ) -> Dict[str, List[Tuple[int, int]]]:
@@ -133,7 +134,11 @@ def get_index_by_type(
     leaves = regions[0].get_leaves()
     if rev:
         leaves.reverse()
-    cuts = get_cuts(leaves)
+    cuts = project_regions_to_coordinates(leaves)
+
+    # index is a legacy data structure, todo fix
+    for c in cuts:
+        index[c.start, c.stop] = c.region_type
 
     # groupby requested region
     for idx, l in enumerate(leaves):
@@ -144,72 +149,31 @@ def get_index_by_type(
     return index
 
 
-def get_index(spec, modality, region_id, rev=False) -> Dict[Tuple[int, int], str]:
+def get_index(
+    spec, modality, region_id, rev=False
+) -> Dict[str, List[RegionCoordinate]]:
     rid = region_id
-    # run function
-    index = defaultdict()
     regions = run_find(spec, modality, rid)
     leaves = regions[0].get_leaves()
     if rev:
         leaves.reverse()
-    cuts = get_cuts(leaves)
+    cuts = project_regions_to_coordinates(leaves)
 
-    for idx, l in enumerate(leaves):
-        t = l.region_type
-        c = cuts[idx]
-        index[c] = t
-
-    return index
+    return {region_id: cuts}
 
 
 def get_index_by_primer(
-    spec, modality: str, read_id: str, rev: bool = False
-) -> Dict[Tuple[int, int], str]:  # noqa
-    index = defaultdict()
-
-    # get all atomic elements from library
-    leaves = spec.get_libspec(modality).get_leaves()
-
-    # get the read object and primer id
-    read = [i for i in spec.sequence_spec if i.read_id == read_id][0]
-    primer_id = read.primer_id
-
-    # get the index of the primer in the list of leaves (ASSUMPTION, 5'->3' and primer is an atomic element)
-    primer_idx = [i for i, l in enumerate(leaves) if l.region_id == primer_id][0]
-
-    # If we are on the opposite strand, we go in the opposite way
-    if read.strand == "neg":
-        rgn = leaves[:primer_idx][::-1]
-    else:
-        rgn = leaves[primer_idx + 1 :]
+    spec, modality: str, read_id: str
+) -> Dict[str, List[RegionCoordinate]]:  # noqa
+    # this manages the strandedness internally
+    (read, rgns) = map_read_id_to_regions(spec, modality, read_id)
 
     # get the cuts for all of the atomic elements (tuples of 0-indexed start stop)
-    cuts = get_cuts(rgn)
+    rcs = project_regions_to_coordinates(rgns)
 
-    # associate each cut with its region type
-    for idx, r in enumerate(rgn):
-        t = r.region_type
-        c = cuts[idx]
-        index[c] = t
-    # intersect read with index
-    new_index = itx_read(index, read.max_len)
+    new_rcs = itx_read(rcs, 0, read.max_len)
 
-    return new_index
-
-
-def itx_read(index: Dict[Tuple[int, int], str], read_stop: int):
-    new_index = defaultdict()
-
-    # three cases
-    for (itv_start, itv_stop), e in index.items():  # noqa
-        # A: read is shorter than itv_start of next element
-        if read_stop < itv_start:
-            continue
-
-        # B: read splits an element (same as case c)
-        elif read_stop > itv_start:
-            new_index[itv_start, min(itv_stop, read_stop)] = e
-    return new_index
+    return {read_id: new_rcs}
 
 
 def format_kallisto_bus(indices, subregion_type=None):
@@ -217,19 +181,19 @@ def format_kallisto_bus(indices, subregion_type=None):
     umi = []
     feature = []
     for idx, region in enumerate(indices):
-        for rgn, index in region.items():
-            for k, v in index.items():
-                if v.upper() == "BARCODE":
-                    bcs.append(f"{idx},{k[0]},{k[1]}")
-                elif v.upper() == "UMI":
-                    umi.append(f"{idx},{k[0]},{k[1]}")
+        for rgn, cuts in region.items():
+            for cut in cuts:
+                if cut.region_type.upper() == "BARCODE":
+                    bcs.append(f"{idx},{cut.start},{cut.stop}")
+                elif cut.region_type.upper() == "UMI":
+                    umi.append(f"{idx},{cut.start},{cut.stop}")
                 elif (
-                    v.upper() == "CDNA"
-                    or v.upper() == "GDNA"
-                    or v.upper() == "PROTEIN"
-                    or v.upper() == "TAG"
+                    cut.region_type.upper() == "CDNA"
+                    or cut.region_type.upper() == "GDNA"
+                    or cut.region_type.upper() == "PROTEIN"
+                    or cut.region_type.upper() == "TAG"
                 ):
-                    feature.append(f"{idx},{k[0]},{k[1]}")
+                    feature.append(f"{idx},{cut.start},{cut.stop}")
     if len(umi) == 0:
         umi.append("-1,-1,-1")
     if len(bcs) == 0:
@@ -240,6 +204,7 @@ def format_kallisto_bus(indices, subregion_type=None):
 
 
 # this one should only return one string
+# TODO: return to this
 def format_seqkit_subseq(indices, subregion_type=None):
     # The x string format is start:stop (1-indexed)
     # x = ""
@@ -247,10 +212,10 @@ def format_seqkit_subseq(indices, subregion_type=None):
     # # need to get the right start position
     x = ""
     region = indices[0]
-    for rgn, index in region.items():
-        for k, v in index.items():
-            if v == subregion_type:
-                x = f"{k[0]+1}:{k[1]}\n"
+    for rgn, cuts in region.items():
+        for cut in cuts:
+            if cut.region_type == subregion_type:
+                x = f"{cut.start+1}:{cut.stop}\n"
 
     return x
 
@@ -258,9 +223,9 @@ def format_seqkit_subseq(indices, subregion_type=None):
 def format_tab(indices, subregion_type=None):
     x = ""
     for idx, region in enumerate(indices):
-        for rgn, index in region.items():
-            for k, v in index.items():
-                x += f"{rgn}\t{v}\t{k[0]}\t{k[1]}\n"
+        for rgn, cuts in region.items():
+            for cut in cuts:
+                x += f"{rgn}\t{cut.name}\t{cut.region_type}\t{cut.start}\t{cut.stop}\n"
 
     return x[:-1]
 
@@ -270,14 +235,16 @@ def format_starsolo(indices, subregion_type=None):
     umi = []
     cdna = []
     for idx, region in enumerate(indices):
-        for rgn, index in region.items():
-            for k, v in index.items():
-                if v.upper() == "BARCODE":
-                    bcs.append(f"--soloCBstart {k[0] + 1} --soloCBlen {k[1]}")
-                elif v.upper() == "UMI":
-                    umi.append(f"--soloUMIstart {k[0] + 1} --soloUMIlen {k[1] - k[0]}")
-                elif v.upper() == "CDNA":
-                    cdna.append(f"{k[0]},{k[1]}")
+        for rgn, cuts in region.items():
+            for cut in cuts:
+                if cut.region_type.upper() == "BARCODE":
+                    bcs.append(f"--soloCBstart {cut.start + 1} --soloCBlen {cut.stop}")
+                elif cut.region_type.upper() == "UMI":
+                    umi.append(
+                        f"--soloUMIstart {cut.start + 1} --soloUMIlen {cut.stop - cut.start}"
+                    )
+                elif cut.region_type.upper() == "CDNA":
+                    cdna.append(f"{cut.start},{cut.stop}")
     x = f"--soloType CB_UMI_Simple {bcs[0]} {umi[0]}"
     return x
 
@@ -288,14 +255,14 @@ def format_simpleaf(indices, subregion_type=None):
     for idx, region in enumerate(indices):
         fn = idx
         x = f"{fn+1}{{"
-        for rgn, index in region.items():
-            for k, v in index.items():
-                if v.upper() == "BARCODE":
-                    x += f"b[{k[1]-k[0]}]"
-                elif v.upper() == "UMI":
-                    x += f"u[{k[1]-k[0]}]"
-                elif v.upper() == "CDNA":
-                    x += f"r[{k[1] - k[0]}]"
+        for rgn, cuts in region.items():
+            for cut in cuts:
+                if cut.region_type.upper() == "BARCODE":
+                    x += f"b[{cut.stop-cut.start}]"
+                elif cut.region_type.upper() == "UMI":
+                    x += f"u[{cut.stop-cut.start}]"
+                elif cut.region_type.upper() == "CDNA":
+                    x += f"r[{cut.stop - cut.start}]"
             x += "x:}"
         xl.append(x)
     return "".join(xl)
@@ -305,14 +272,14 @@ def format_zumis(indices, subregion_type=None):
     xl = []
     for idx, region in enumerate(indices):
         x = ""
-        for rgn, index in region.items():
-            for k, v in index.items():
-                if v.upper() == "BARCODE":
-                    x += f"- BCS({k[0] + 1}-{k[1]})\n"
-                elif v.upper() == "UMI":
-                    x += f"- UMI({k[0] + 1}-{k[1]})\n"
-                elif v.upper() == "CDNA":
-                    x += f"- cDNA({k[0] + 1}-{k[1]})\n"
+        for rgn, cuts in region.items():
+            for cut in cuts:
+                if cut.region_type.upper() == "BARCODE":
+                    x += f"- BCS({cut.start + 1}-{cut.stop})\n"
+                elif cut.region_type.upper() == "UMI":
+                    x += f"- UMI({cut.start + 1}-{cut.stop})\n"
+                elif cut.region_type.upper() == "CDNA":
+                    x += f"- cDNA({cut.start + 1}-{cut.stop})\n"
         xl.append(x)
 
     return "\n".join(xl)[:-1]
@@ -324,19 +291,19 @@ def format_chromap(indices, subregion_type=None):
     gdna_fqs = []
     gdna_str = []
     for idx, region in enumerate(indices):
-        for rgn, index in region.items():
-            for k, v in index.items():
-                if v.upper() == "BARCODE":
+        for rgn, cuts in region.items():
+            for cut in cuts:
+                if cut.region_type.upper() == "BARCODE":
                     bc_fqs.append(rgn)
-                    bc_str.append(f"bc:{k[0]}:{k[1]}")
+                    bc_str.append(f"bc:{cut.start}:{cut.stop}")
                     pass
-                elif v.upper() == "GDNA":
+                elif cut.region_type.upper() == "GDNA":
                     gdna_fqs.append(rgn)
-                    gdna_str.append(f"{k[0]}:{k[1]}")
+                    gdna_str.append(f"{cut.start}:{cut.stop}")
     if len(set(bc_fqs)) > 1:
-        raise "chromap only supports barcodes from one fastq"
+        raise Exception("chromap only supports barcodes from one fastq")
     if len(set(gdna_fqs)) > 2:
-        raise "chromap only supports genomic dna from two fastqs"
+        raise Exception("chromap only supports genomic dna from two fastqs")
 
     barcode_fq = bc_fqs[0]
     read1_fq = list(set(gdna_fqs))[0]