+++ seqspec/utils.py
@@ -28,16 +28,21 @@ def load_genbank_stream(gbk_stream: io.IOBase):
     return data
 
 
-# return cut indices for all atomic regions
-def get_cuts(regions, cuts=[]):
-    if not cuts:
-        cuts = []
-    prev = 0
-    for r in regions:
-        nxt = prev + r.max_len
-        cuts.append((prev, nxt))
-        prev = nxt
-    return cuts
+class RegionCoordinate:
+    def __init__(self, cut_name, cut_type, start, end):
+        self.cut_name = cut_name
+        self.cut_type = cut_type
+        self.start = start
+        self.end = end
+
+    def __repr__(self):
+        return f"RegionCoordinate {self.cut_name} [{self.cut_type}]: ({self.start}, {self.end})"
+
+    def __str__(self):
+        return f"RegionCoordinate {self.cut_name} [{self.cut_type}]: ({self.start}, {self.end})"
+
+    def __eq__(self, other):
+        return self.start == other.start and self.end == other.end
 
 
 def write_read(header, seq, qual, f):
@@ -126,3 +131,20 @@ def complement_nucleotide(nucleotide):
 
 def complement_sequence(sequence):
     return "".join(complement_nucleotide(n) for n in sequence.upper())
+
+
+def map_read_id_to_regions(spec, modality, region_id):
+    # get all atomic elements from library
+    leaves = spec.get_libspec(modality).get_leaves()
+    # get the read object and primer id
+    read = [i for i in spec.sequence_spec if i.read_id == region_id][0]
+    primer_id = read.primer_id
+    # get the index of the primer in the list of leaves (ASSUMPTION, 5'->3' and primer is an atomic element)
+    primer_idx = [i for i, l in enumerate(leaves) if l.region_id == primer_id][0]
+    # If we are on the opposite strand, we go in the opposite way
+    if read.strand == "neg":
+        rgns = leaves[:primer_idx][::-1]
+    else:
+        rgns = leaves[primer_idx + 1 :]
+
+    return (read, rgns)